refactor(agent): 重构 AgentController，合并对话与历史接口
- 合并原 ChatController 至 AgentController，形成统一入口 API 体系
- 新增同步对话和流式对话接口，支持实时和非实时交互
- 新增对话历史查询和清除接口，便于管理会话数据
- 使用 SseExecutor 统一封装 SSE 心跳保活、断连检测和清理逻辑
- 移除过时虚拟线程及心跳调度器相关代码，简化 SSE 流式处理
- 更新请求守卫，支持幂等、限流及审计日志管理
- 调整 AgentAutoConfiguration，增加 PromptTemplateProvider 注入
- 精简 AgentFeign 启用配置，默认开启自动扫描包路径
- BatchScheduler 改为指标采集器，仅采集系统负载指标并暴露，不再调整 batchSize
- 更新配置元数据，新增异常恢复相关参数和名称规范调整
- 调整文档说明，明确 AgentExecutor 新增执行器模式及 RouterAgentExecutor 已移除

+++ ydsz-agent/ydsz-agent-server/src/main/java/com/njydsz/agent/server/analytics/CostAnalysisService.java
@@ -70,9 +70,9 @@ public class CostAnalysisService {
           TokenUsageRecordDO.builder()
               .conversationId(conversationId)
               .modelName(modelName)
-              .promptTokens(usage.getPromptTokens())
-              .completionTokens(usage.getCompletionTokens())
-              .totalTokens(usage.getTotalTokens())
+              .promptTokens((long) usage.getPromptTokens())
+              .completionTokens((long) usage.getCompletionTokens())
+              .totalTokens((long) usage.getTotalTokens())
               .build();
       usageRecordMapper.insert(record);
     } catch (Exception e) {
